fixing warnings refs #12
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@21690 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/MSLeaderInfo.cpp
@@ -201,7 +201,7 @@ MSLeaderDistanceInfo::addLeader(const MSVehicle* veh, SUMOReal gap, SUMOReal lat
         // speedup for the simple case
         sublane = 0;
     }
-    if (sublane >= 0 && sublane < myVehicles.size()) {
+    if (sublane >= 0 && sublane < (int)myVehicles.size()) {
         // sublane is already given
         if (gap < myDistances[sublane]) {
             if (myVehicles[sublane] == 0) {
@@ -307,7 +307,7 @@ MSCriticalFollowerDistanceInfo::addFollower(const MSVehicle* veh, const MSVehicl
         // speedup for the simple case
         sublane = 0;
     }
-    if (sublane >= 0 && sublane < myVehicles.size()) {
+    if (sublane >= 0 && sublane < (int)myVehicles.size()) {
         // sublane is already given
         if (missingGap > myMissingGaps[sublane]) {
             if (myVehicles[sublane] == 0) {
